Add network config daemon
This is a Python script that spawns a daemon, which is capable of
broadcasting the network configuration of the machine via UDP.

The intended use case is that a second machine will send out a UDP
request, causing the daemon on the first machine to broadcast its
configuration.  The second machine can then use this information to
connect to the first.  As such, it can be used to establish network
connections to machines which do not have static IP addresses.